More modernization.

+++ NSPopUpButtonExtended.h
@@ -10,7 +10,7 @@
 
 @interface NSPopUpButton (Extended)
 
--(id) addItemWithTitle:(NSString*)title tag:(int)tag;
--(NSMenuItem*) addItemWithTitle:(NSString*)aString target:(id)target action:(SEL)aSelector tag:(int)tag;
+-(id) addItemWithTitle:(NSString*)title tag:(NSInteger)tag;
+-(NSMenuItem*) addItemWithTitle:(NSString*)aString target:(id)target action:(SEL)aSelector tag:(NSInteger)tag;
 
 @end
